requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
boto3==1.34.0
pydantic[email]==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
httpx==0.25.2
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
moto[dynamodb]==4.2.12
python-dotenv==1.0.0
cachetools==5.3.2
